+++ R/JSON.R
@@ -10,7 +10,7 @@
 #'   label = label,
 #'   column = column,
 #'   param = param,
-#'   value=value) %>%
+#'   value=value) |>
 #'   tfrmt_to_json()
 #'
 #' @importFrom jsonlite toJSON validate
@@ -132,7 +132,7 @@ as_json.frmt <- function(x){
 
 #' @export
 as_json.frmt_when <- function(x){
-  lhs <- map_chr(x$frmt_ls, f_lhs)
+  lhs <- map_chr(x$frmt_ls, f_lhs_as_char)
   rhs <- map(x$frmt_ls, f_rhs) %>% map(as_json)
   names(rhs) <- lhs
   list(frmt_when = list(frmt_ls = rhs, missing = x$missing))