+++ R/body_plan.R
@@ -59,7 +59,6 @@ body_plan <- function(...){
 #' @importFrom purrr map_dfr map map_chr quietly pmap_chr
 #' @importFrom dplyr mutate group_by filter group_split select across
 #' @importFrom tidyr unnest
-#' @importFrom tidyselect everything
 #' @importFrom rlang as_name quo_is_missing
 body_plan_builder <- function(data, group, label, param_defaults, missing = NULL){
 
@@ -75,7 +74,9 @@ body_plan_builder <- function(data, group, label, param_defaults, missing = NULL
            single_glue_to_frmt = pmap_chr(list(.data$contains_glue, .data$param, .data$param_display), function(a,b,c){
              if(a==TRUE & length(b) == 1) c else NA_character_
            } )) %>%
-    unnest(everything()) %>%
+    unnest(
+      tidyselect::everything()
+    ) %>%
     mutate(frmt_string = map2_chr(.data$sigdig, .data$single_glue_to_frmt, sigdig_frmt_string))
 
   frmt_vec <- param_tbl %>%