+++ R/col_plan.R
@@ -92,6 +92,16 @@
 #'    )
 #'  )
 #'
+#' ## To add a stub header rename the group variable in the column plan
+#' ## If multiple group variables exist, any of them can be renamed.
+#' ## If more than one is renamed, {tfrmt} will use the highest level group name available.
+#'
+#'  renaming_group <- col_plan(
+#'     my_grp = group, # rename group
+#'     label,
+#'     starts_with("col")
+#'   )
+#'
 ##' @section Images:
 #' Here are some example outputs:
 #'
@@ -189,7 +199,6 @@ is_valid_span_structure_call <- function(x){
   as.character(as.list(x)[[1]]) %in% c("span_structure")
 }
 
-#' @importFrom tidyselect vars_select_helpers
 is_valid_tidyselect_call <- function(x){
   ## drop - from determining if
   if(as.character(as.list(x)[[1]]) == "-"){
@@ -198,7 +207,7 @@ is_valid_tidyselect_call <- function(x){
       return(TRUE)
     }
   }
-  as.character(as.list(x)[[1]]) %in% c(names(vars_select_helpers))
+  as.character(as.list(x)[[1]]) %in% c(names(tidyselect::vars_select_helpers))
 }
 
 is_valid_quo_call <- function(x){
@@ -244,4 +253,3 @@ check_col_plan_dots <- function(x){
     }
   })
 }
-