+++ NEWS.md
@@ -1,3 +1,5 @@
+# tfrmt (development version)
+
 # tfrmt 0.2.0
 
 ## Improvements