+++ NEWS.md
@@ -1,3 +1,5 @@
+# tfrmt development version
+
 # tfrmt 0.2.1
 
 Patch release for latest {gt} release and upcoming {purrr} and {stringr} releases: