+++ NEWS.md
@@ -1,5 +1,8 @@
 # tfrmt development version
 
+## Improvements
+ * Replace all spaces in the cell body of an output gt with unicode whitespace to ensure formatting is retained when going to pdf
+
 # tfrmt 0.2.1
 
 Patch release for latest {gt} release and upcoming {purrr} and {stringr} releases: